added a base BNN class and fixed MCMC sampler bug in pybnn repo
